% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_tree.R
\name{champion_tree}
\alias{champion_tree}
\title{Create a champion tree object}
\usage{
champion_tree(
  circumference,
  height,
  spread,
  common = "unknown",
  scientific = "unknown"
)
}
\arguments{
\item{circumference}{The circumference of the tree, in inches}

\item{height}{The height of the tree, in feet}

\item{spread}{The spread of the tree's crown, in feet}

\item{common}{The tree's common name}

\item{scientific}{The tree's scientific name}
}
\value{
An object of class "championtree"
}
\description{
Create a champion tree object
}
\examples{
 champion_tree(1020, 274, 107)
}
