% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_helpers.R
\name{crown_helpers}
\alias{crown_helpers}
\alias{crown_helper}
\alias{crown_helper_faster}
\alias{crown_helper_cpp}
\title{Helper functions for crown}
\usage{
crown_helper(points, species)

crown_helper_faster(points, species)

crown_helper_cpp(points, species)
}
\arguments{
\item{points}{The number of points for each tree}

\item{species}{The species of each tree}
}
\description{
Helper functions for crown
}
\details{
Note that the vectors should be sorted by species and descending points
}
