% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\alias{get_points.champion_tree}
\title{Extract points}
\usage{
get_points(x, ...)

\method{get_points}{champion_tree}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Other arguments}
}
\description{
Extract points
}
