% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\name{Ops.champion_tree}
\alias{Ops.champion_tree}
\title{Compare champion tree objects}
\usage{
\method{Ops}{champion_tree}(e1, e2)
}
\arguments{
\item{e1, e2}{champion_tree objects, or numbers to compare them to.}
}
\description{
Compare champion tree objects
}
